+++ Projects/LastDance/LastDance/Sources/Data/Network/Visitor/VisitorAPIService.swift
@@ -7,7 +7,6 @@
 
 import Foundation
 import Moya
-import UIKit
 
 // MARK: VisitorAPIServiceProtocol
 
@@ -73,8 +72,6 @@ final class VisitorAPIService: VisitorAPIServiceProtocol {
                             let model = VisitorMapper.toModel(from: dto)
                             SwiftDataManager.shared.upsertVisitor(model)
                         }
-                        // TODO: - 전체 Visitors 확인 용도 (이후에 제거 가능)
-                        //                        SwiftDataManager.shared.printAllVisitors()
                     }
 
                     completion(.success(items))
@@ -115,11 +112,6 @@ final class VisitorAPIService: VisitorAPIServiceProtocol {
                     )
 
                     UserDefaults.standard.set(dto.id, forKey: UserDefaultsKey.visitorId.rawValue)
-
-                    // visitor 생성 후 디바이스 토큰 전송
-                    if let appDelegate = UIApplication.shared.delegate as? AppDelegate {
-                        appDelegate.registerDeviceTokenIfNeeded()
-                    }
 
                     completion(.success(dto))
                 } catch {
