+++ Projects/LastDance/LastDance/Sources/Presentation/Onboarding/ViewModel/IdentitySelectionViewModel.swift
@@ -79,6 +79,9 @@ final class IdentitySelectionViewModel: ObservableObject {
                         name: dto.name
                     )
                     self.dataManager.insert(visitor)
+
+                    // visitor 생성 성공 후 디바이스 토큰 전송
+                    DeviceTokenManager.shared.registerDeviceTokenIfNeeded()
                 case .failure(let error):
                     if let moyaError = error as? MoyaError,
                         let data = moyaError.response?.data,
